+++ km/com/hifiremote/jp1/EncrypterDecrypter.java
@@ -2,6 +2,6 @@ package com.hifiremote.jp1;
 
 public abstract class EncrypterDecrypter
 {
-  public abstract int encrypt( int val );
-  public abstract int decrypt( int val );
+  public abstract short encrypt( short val );
+  public abstract short decrypt( short val );
 }
